requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-inference==1.0.0b9
azure-core==1.35.0
azure-identity==1.23.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.5
discord.py==2.5.2
fastapi==0.116.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
isodate==0.7.2
msal==1.32.3
msal-extensions==1.3.1
multidict==6.6.3
propcache==0.3.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.1
requests==2.32.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
yarl==1.20.1
